+++ gems/aws-sdk-core/spec/seahorse/client/base_spec.rb
@@ -274,7 +274,7 @@ module Seahorse
             Plugins::NetHttp,
             Plugins::RaiseResponseErrors,
             Plugins::ResponseTarget,
-            Plugins::ChunkSentCallback
+            Plugins::RequestCallback
           ])
         end